feat: folder structuring
- rm unused deps
- rm unused files & declarations
- cleanup
