+++ app/_components/About/Technologies.tsx
