+++ app/_components/Services.tsx
@@ -1,8 +1,8 @@
 "use client";
-import Section from "../base/Section";
 import { AnimatePresence, motion } from "motion/react";
 import { Service } from "@/lib/types";
 import { IconBulb, IconBugOff, IconCode, IconTools } from "@tabler/icons-react";
+import Section from "@/components/layouts/Section";
 
 const services: Service[] = [
   {
