+++ app/projects/_components/ProjectItem.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { motion } from "motion/react";
 import { TProject } from "@/lib/types";
 import { IconArrowUpRight, IconBrandGithub } from "@tabler/icons-react";
-import Tag from "../base/Tag";
+import Tag from "@/components/core/Tag";
 
 type Props = {
   project: TProject;
