+++ lib/fonts.ts
@@ -6,17 +6,3 @@ export const rubik = Rubik({
   display: "swap",
   variable: "--font-serif",
 });
-
-export const inconsolata = Inconsolata({
-  weight: ["400", "500", "600", "700"],
-  subsets: ["latin"],
-  display: "swap",
-  variable: "--font-mono",
-});
-
-export const oleoScript = Oleo_Script({
-  weight: ["400", "700"],
-  subsets: ["latin"],
-  display: "swap",
-  variable: "--font-display",
-});
