--- lib/utils/math.ts
@@ -1,14 +1,0 @@
-export function getEllipseY(
-  x: number,
-  h: number,
-  k: number,
-  a: number,
-  b: number,
-) {
-  const term = a * a - (x - h) * (x - h);
-  if (term < 0) return 0;
-
-  const sqrt = Math.sqrt(term);
-
-  return k + (b * sqrt) / a;
-}
