+++ package.json
@@ -39,7 +39,6 @@
     "motion": "^12.23.12",
     "next": "15.4.6",
     "react": "19.1.1",
-    "react-circular-progressbar": "^2.1.0",
     "react-dom": "19.1.1",
     "react-fast-marquee": "^1.6.5",
     "tailwind-merge": "^2.6.0"
@@ -56,6 +55,7 @@
     "eslint-plugin-prettier": "^5.2.1",
     "husky": "^9.1.7",
     "postcss": "^8.5.6",
+    "postcss-load-config": "^6.0.1",
     "prettier": "^3.3.3",
     "prettier-eslint": "^16.3.0",
     "prettier-plugin-tailwindcss": "^0.6.5",
