+++ utils/string.ts
@@ -1,0 +1,11 @@
+export const toKebabCase = (str: string | null | undefined): string => {
+  if (!str) return "";
+  return (
+    str
+      .match(
+        /[A-Z]{2,}(?=[A-Z][a-z]+[0-9]*|\b)|[A-Z]?[a-z]+[0-9]*|[A-Z]|[0-9]+/g,
+      )
+      ?.map((x) => x.toLowerCase())
+      .join("-") ?? ""
+  );
+};
